Fix Lua types and update Stopwatch script

+++ src/hal/time.cpp
@@ -39,8 +39,8 @@ void OswHal::setupTime(void) {
 
 bool OswHal::hasDS3231(void) { return getUTCTime() > 0; }
 
-long OswHal::getUTCTime(void) { return Rtc.GetDateTime().Epoch32Time(); }
-long OswHal::getLocalTime(void) { return getUTCTime() + 3600 * TIMEZONE + 3600 * DAYLIGHTOFFSET; }
+uint32_t OswHal::getUTCTime(void) { return Rtc.GetDateTime().Epoch32Time(); }
+uint32_t OswHal::getLocalTime(void) { return getUTCTime() + 3600 * TIMEZONE + 3600 * DAYLIGHTOFFSET; }
 void OswHal::setUTCTime(long epoch) {
   RtcDateTime t = RtcDateTime();
   t.InitWithEpoch32Time(epoch);
